[Patchset 6] ftrace_reader: Update for  uint64 block device ids
Switch previous code for plumbing uint32 device ids to translate to uint64 and add
translation strategies for uint64 device ids.

Bug: 73873362
Bug: 73625480

+++ src/ftrace_reader/cpu_reader.h
@@ -115,7 +115,7 @@ class CpuReader {
                         protozero::Message* out,
                         FtraceMetadata* metadata) {
     T t = ReadIntoVarInt<T>(start, field_id, out);
-    metadata->AddInode(t);
+    metadata->AddInode(static_cast<uint64_t>(t));
   }
 
   template <typename T>
@@ -125,7 +125,7 @@ class CpuReader {
                         FtraceMetadata* metadata) {
     T t = ReadIntoVarInt<T>(start, field_id, out);
     PERFETTO_DCHECK(t != 0);
-    metadata->AddDevice(t);
+    metadata->AddDevice(static_cast<uint64_t>(t));
   }
 
   static void ReadPid(const uint8_t* start,
